pyproject.toml
[tool.poetry]
name = "habittrackerapi"
version = "0.1.0"
description = ""
authors = ["Kirill Tuchkin <volkirtu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
django-celery-beat = "^2.5.0"
celery = "^5.3.4"
django-cors-headers = "^4.2.0"
pytelegrambotapi = "^4.14.0"
redis = "^5.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
